+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [v2.50.1](https://github.com/docling-project/docling-core/releases/tag/v2.50.1) - 2025-11-04
+
+### Fix
+
+* Add JSON to CodeLanguageLabel ([#413](https://github.com/docling-project/docling-core/issues/413)) ([`09ef91c`](https://github.com/docling-project/docling-core/commit/09ef91c27285f3bdd4f41c935fd49685ff26524b))
+
+## [v2.50.0](https://github.com/docling-project/docling-core/releases/tag/v2.50.0) - 2025-10-30
+
+### Feature
+
+* Add metadata model hierarchy ([#408](https://github.com/docling-project/docling-core/issues/408)) ([`2ee3cac`](https://github.com/docling-project/docling-core/commit/2ee3cacdd640414a3e9a8b861a15c4baba6f1d06))
+* Add split view & YAML support to CLI viewer ([#407](https://github.com/docling-project/docling-core/issues/407)) ([`a3feae0`](https://github.com/docling-project/docling-core/commit/a3feae0a1e4b8adfe05009f7a12e648e8aedd149))
+* New picture classes for doctags ([#404](https://github.com/docling-project/docling-core/issues/404)) ([`ada4068`](https://github.com/docling-project/docling-core/commit/ada4068e7a1b3b32e6dcda93658fae8b98ddfb4d))
+
+## [v2.49.0](https://github.com/docling-project/docling-core/releases/tag/v2.49.0) - 2025-10-16
+
+### Feature
+
+* Python 3.14 compatibility ([#403](https://github.com/docling-project/docling-core/issues/403)) ([`47b70d3`](https://github.com/docling-project/docling-core/commit/47b70d3c246f7aa8fe4967cbad07660c15d0b505))
+* Rendering of checkboxes in markdown ([#381](https://github.com/docling-project/docling-core/issues/381)) ([`abfa81f`](https://github.com/docling-project/docling-core/commit/abfa81f0651e70c1b3bc479f5621f706d66f0cbd))
+
 ## [v2.48.4](https://github.com/docling-project/docling-core/releases/tag/v2.48.4) - 2025-10-01
 
 ### Fix